// src/interface.tsx
// export default interface Brewery {
//   id: number;
//   name: string;
//   brewery_type: string;
//   street: string;
//   city: string;
//   state: string;
//   postal_code: string;
//   country: string;
//   longitude: string;
//   latitude: string;
//   phone: string;
//   website_url: string;
//   avatarColor: string; 
// }


export default interface Brewery {
  id: number;
  name: string;
  brewery_type: string;
  address_1: string;
  address_2: string | null;
  address_3: string | null;
  city: string;
  state_province: string;
  postal_code: string;
  country: string;
  longitude: string;
  latitude: string;
  phone: string;
  website_url: string;
  state: string;
  street: string;
  avatarColor: string;
}